Room的增删改查的修改

+++ app/src/main/java/com/zhihaoliang/roomtest/room/pair/KeyAndValueDao.java
@@ -29,6 +29,6 @@ public interface KeyAndValueDao {
     int update(KeyAndValue keyAndValue);
 
     @Delete
-    int delete(KeyAndValue keyAndValue);
+    long delete(KeyAndValue keyAndValue);
 
 }
